+++ Tests/SentryTests/Profiling/SentryProfilerTests.mm
@@ -33,6 +33,22 @@
         @"-[SentryProfilerTests testParseFunctionNameWithBacktraceSymbolsInput]");
 }
 
+- (void)testProfilerCanBeInitializedOnMainThread
+{
+    XCTAssertNotNil([[SentryProfiler alloc] init]);
+}
+
+- (void)testProfilerCannotBeInitializedOffMainThread
+{
+    const auto expectation = [self expectationWithDescription:@"background initializing profiler"];
+    dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0ul), ^{
+        XCTAssertNil([[SentryProfiler alloc] init]);
+        [expectation fulfill];
+    });
+    [self waitForExpectationsWithTimeout:1.0
+                                 handler:^(NSError *_Nullable error) { NSLog(@"%@", error); }];
+}
+
 @end
 
 #endif